detail 폴더 변경
